requirements/base.txt
Flask==1.1.2
Flask-Caching==1.9.0
Flask-Cors==3.0.9
flask-marshmallow==0.13.0
flask-smorest==0.24.1
Flask-SQLAlchemy==2.4.4
pyodbc==4.0.30
gunicorn==20.0.4
maaskeycloak==0.3.0
maaslogger==0.4.0
marshmallow==3.8.0
python-dotenv==0.14.0
redis==3.5.3
requests==2.24.0